requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
pydantic
pymysql
alembic
python-dotenv
httpx
